pyproject.toml
[project]
name = "flashcard-llm"
version = "0.1.0"
description = "Generate flashcards CSV from PDF/text using DeepSeek via LangChain"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "langchain>=0.2.16",
  "langchain-openai>=0.1.14",
  "langchain-text-splitters>=0.3.9",
  "pypdf>=3.17.4",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.4",
]

 
